MAINT: Relocate all schema classes together
