+++ src/fmu/dataio/_models/products/inplace_volumes.py
@@ -1,18 +1,15 @@
 from __future__ import annotations
 
 from pathlib import Path
-from typing import TYPE_CHECKING, Dict, List, Literal, Optional, TypeVar
+from typing import TYPE_CHECKING, List, Optional
 
 from pydantic import BaseModel, Field, RootModel
-from pydantic.json_schema import GenerateJsonSchema
 
-from fmu.dataio._definitions import FmuSchemas, SchemaBase
+from fmu.dataio._models._schema_base import FmuSchemas, SchemaBase
 from fmu.dataio.export._enums import InplaceVolumes
 
 if TYPE_CHECKING:
-    from typing import Any, Mapping
-
-T = TypeVar("T", Dict, List, object)
+    from typing import Any
 
 
 class InplaceVolumesResultRow(BaseModel):
@@ -53,43 +50,8 @@ class InplaceVolumesSchema(SchemaBase):
     FILENAME: str = "inplace_volumes.json"
     PATH: Path = FmuSchemas.PATH / "file_formats" / VERSION / FILENAME
 
-    class InplaceVolumesGenerateJsonSchema(GenerateJsonSchema):
-        """Implements a schema generator so that some additional fields may be added."""
-
-        def _remove_format_path(self, obj: T) -> T:
-            """
-            Removes entries with key "format" and value "path" from dictionaries. This
-            adjustment is necessary because JSON Schema does not recognize the "format":
-            "path", while OpenAPI does. This function is used in contexts where OpenAPI
-            specifications are not applicable.
-            """
-
-            if isinstance(obj, dict):
-                return {
-                    k: self._remove_format_path(v)
-                    for k, v in obj.items()
-                    if not (k == "format" and v == "path")
-                }
-
-            if isinstance(obj, list):
-                return [self._remove_format_path(element) for element in obj]
-
-            return obj
-
-        def generate(
-            self,
-            schema: Mapping[str, Any],
-            mode: Literal["validation", "serialization"] = "validation",
-        ) -> dict[str, Any]:
-            json_schema = super().generate(schema, mode=mode)
-            json_schema["$schema"] = self.schema_dialect
-            json_schema["$id"] = InplaceVolumesSchema.url()
-            json_schema["version"] = InplaceVolumesSchema.VERSION
-
-            return json_schema
-
-    @staticmethod
-    def dump() -> dict[str, Any]:
+    @classmethod
+    def dump(cls) -> dict[str, Any]:
         return InplaceVolumesResult.model_json_schema(
-            schema_generator=InplaceVolumesSchema.InplaceVolumesGenerateJsonSchema
+            schema_generator=cls.default_generator()
         )
